feat: move server side fetch helper function to server utils file
